Extracted small_circular_button
